ariel_os_stm32::peripherals

Struct I2C1

pub struct I2C1 { /* private fields */ }

I2C1 peripheral

Implementations

impl I2C1

pub unsafe fn steal() -> Peri<'static, I2C1>

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.