ariel_os_stm32::peripherals

Struct LPTIM2

pub struct LPTIM2 { /* private fields */ }

LPTIM2 peripheral

Implementations

impl LPTIM2

pub unsafe fn steal() -> Peri<'static, LPTIM2>

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.