ariel_os_stm32::peripherals

Struct PA2

pub struct PA2 { /* private fields */ }

PA2 peripheral

Implementations

impl PA2

pub unsafe fn steal() -> PA2

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.