ariel_os_stm32::peripherals
Struct PB4

pub struct PB4 { /* private fields */ }

PB4 peripheral

Implementations

impl PB4

pub unsafe fn steal() -> PB4

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.